user connected check jwt put-post discount+product and delete

+++ src/main/java/com/ecommerce/commercial/controller/ProductController.java
@@ -2,6 +2,8 @@ package com.ecommerce.commercial.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,8 +11,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.ecommerce.commercial.config.JwtConfig;
 import com.ecommerce.commercial.model.Product;
 import com.ecommerce.commercial.service.ProductService;
+
+import jakarta.servlet.http.HttpServletRequest;
 
 @RestController
 @CrossOrigin
@@ -35,8 +40,26 @@ public class ProductController {
 //     productService.updateProduct(id, product);
 //   }
 
+// @DeleteMapping("{id}")
+//   public void deleteProduct(@PathVariable(name="id") Long id) {
+//     productService.deleteProduct(id);
+// }
+
 @DeleteMapping("{id}")
-  public void deleteProduct(@PathVariable(name="id") Long id) {
+public ResponseEntity<String> deleteProduct(@PathVariable(name="id") Long id, HttpServletRequest request) {
+    String jwt = request.getHeader("Authorization");
+    
+    if (jwt == null || jwt.isEmpty()) {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+    }
+    String newJwt = jwt.replace("Bearer ", "");
+    System.out.print(newJwt);
+    if (JwtConfig.validateJwt(newJwt)) {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+    }
+    
     productService.deleteProduct(id);
+    return ResponseEntity.ok("Produit supprimé avec succès");
 }
+
 }
